# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LifeOps personal finance API
 * OpenAPI spec version: 0.3.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface User {
  id: number;
  fullName: string;
  mobile: string;
  token?: string;
  createdAt: string;
  updatedAt: string;
}

export interface RegisterRequest {
  fullName: string;
  mobile: string;
  password: string;
  confirmPassword: string;
}

export interface LoginRequest {
  mobile: string;
  password: string;
}

export interface ChangePasswordRequest {
  currentPassword: string;
  newPassword: string;
  confirmPassword: string;
}

export interface ForgotPasswordRequest {
  mobile: string;
}

export type CategoryType = typeof CategoryType[keyof typeof CategoryType];


export const CategoryType = {
  income: 'income',
  expense: 'expense',
  loan: 'loan',
  goal: 'goal',
} as const;

export interface Category {
  id: number;
  userId: number;
  name: string;
  type: CategoryType;
  isDefault: boolean;
  icon?: string | null;
  color?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateCategoryRequestType = typeof CreateCategoryRequestType[keyof typeof CreateCategoryRequestType];


export const CreateCategoryRequestType = {
  income: 'income',
  expense: 'expense',
  loan: 'loan',
  goal: 'goal',
} as const;

export interface CreateCategoryRequest {
  name: string;
  type: CreateCategoryRequestType;
}

export interface UpdateCategoryRequest {
  name: string;
  icon?: string;
  color?: string;
}

export interface Income {
  id: number;
  userId: number;
  categoryId: number;
  categoryName?: string;
  amount: number;
  date: string;
  notes?: string;
  recurrenceType: string;
  occurrences: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateIncomeRequest {
  categoryId: number;
  amount: number;
  date: string;
  notes?: string;
  recurrenceType?: string;
  occurrences?: number;
}

export interface Expense {
  id: number;
  userId: number;
  categoryId: number;
  categoryName?: string;
  amount: number;
  date: string;
  notes?: string;
  recurrenceType: string;
  occurrences: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateExpenseRequest {
  categoryId: number;
  amount: number;
  date: string;
  notes?: string;
  recurrenceType?: string;
  occurrences?: number;
}

export interface Loan {
  id: number;
  userId: number;
  name: string;
  loanType: string;
  principalAmount: number;
  interestRate: number;
  tenureYears: number;
  emi: number;
  startDate: string;
  endDate: string;
  outstandingBalance: number;
  principalPaid: number;
  interestPaid: number;
  totalInterest: number;
  monthsCompleted: number;
  monthsRemaining: number;
  totalMonths: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateLoanRequest {
  name: string;
  loanType: string;
  principalAmount: number;
  interestRate: number;
  tenureYears: number;
  startDate: string;
  emi?: number;
}

export interface Goal {
  id: number;
  userId: number;
  name: string;
  targetAmount: number;
  currentAmount: number;
  targetDate: string;
  progressPercentage: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateGoalRequest {
  name: string;
  targetAmount: number;
  currentAmount: number;
  targetDate: string;
}

export type UpcomingRecurringType = typeof UpcomingRecurringType[keyof typeof UpcomingRecurringType];


export const UpcomingRecurringType = {
  income: 'income',
  expense: 'expense',
} as const;

export interface UpcomingRecurring {
  id: number;
  type: UpcomingRecurringType;
  categoryName: string;
  amount: number;
  recurrenceType: string;
  nextDate: string;
}

export interface DashboardSummary {
  totalIncome: number;
  totalExpenses: number;
  savings: number;
  activeLoans: number;
  totalOutstanding: number;
  emiDueThisMonth: number;
  goalsCount: number;
  avgGoalProgress: number;
  monthlyIncomes: Income[];
  monthlyExpenses: Expense[];
  upcomingRecurring: UpcomingRecurring[];
}

export type GetCategoriesParams = {
type?: GetCategoriesType;
};

export type GetCategoriesType = typeof GetCategoriesType[keyof typeof GetCategoriesType];


export const GetCategoriesType = {
  income: 'income',
  expense: 'expense',
} as const;

export type GetIncomesParams = {
month?: string;
};

export type GetExpensesParams = {
month?: string;
};

export type GetDashboardParams = {
month?: string;
};