# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LifeOps personal finance API
 * OpenAPI spec version: 0.3.0
 */

export interface Expense {
  id: number;
  userId: number;
  categoryId: number;
  categoryName?: string;
  amount: number;
  date: string;
  notes?: string;
  recurrenceType: string;
  occurrences: number;
  createdAt: string;
  updatedAt: string;
}