# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LifeOps personal finance API
 * OpenAPI spec version: 0.3.0
 */

export interface User {
  id: number;
  fullName: string;
  mobile: string;
  token?: string;
  createdAt: string;
  updatedAt: string;
}